fix: prop dropdown  and style same css
